+++ app/build.gradle
@@ -23,11 +23,11 @@ android {
     buildTypes {
         def API_URL = "API_URL"
 
-        debug{
-            buildConfigField "String", API_URL,  '"https://swapi.co/api/"'
+        debug {
+            buildConfigField "String", API_URL, '"https://swapi.co/api/"'
         }
         release {
-            buildConfigField "String", API_URL,  '"https://swapi.co/api/"'
+            buildConfigField "String", API_URL, '"https://swapi.co/api/"'
 
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
@@ -41,7 +41,7 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation"org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
 
     // AndroidX Components
     implementation 'androidx.appcompat:appcompat:1.1.0-rc01'
@@ -73,7 +73,7 @@ dependencies {
 
     // ViewModel and LiveData
     implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
-    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0-alpha01"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycle_version"
     implementation "androidx.lifecycle:lifecycle-viewmodel-savedstate:1.0.0-alpha01"
     kapt "androidx.lifecycle:lifecycle-compiler:$lifecycle_version"
 
@@ -87,4 +87,6 @@ dependencies {
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'androidx.test:runner:1.3.0-alpha01'
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.3.0-alpha01'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.1.0'
+    implementation 'androidx.navigation:navigation-ui-ktx:2.1.0'
 }
